// contracts/generated/typescript/models/PageRedirect.ts
export interface PageRedirect {
    /**
     *
     * @type {string}
     * @memberof PageRedirect
     */
    sourcePageId: string;
    /**
     *
     * @type {PageRedirectTarget}
     * @memberof PageRedirect
     */
    target: PageRedirectTarget;
    /**
     *
     * @type {string}
     * @memberof PageRedirect
     */
    createdBy: string;
    /**
     *
     * @type {string}
     * @memberof PageRedirect
     */
    updatedBy: string;
    /**
     *
     * @type {Date}
     * @memberof PageRedirect
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof PageRedirect
     */
    updatedAt: Date;
}